Metodo getDateDTOFromDB implementado. Metodo getDateOFromDateVAU: getOYear implementado

+++ src/main/java/CCASolutions/Calendario/Services/DaysService.java
@@ -4,5 +4,5 @@ import CCASolutions.Calendario.DTOs.DateDTOFromDB;
 
 public interface DaysService {
 	
-	public abstract int getDiasASumarALaLunaNueva(DateDTOFromDB dateVAU);
+	public abstract long getDiasASumarALaLunaNueva(DateDTOFromDB dateVAU);
 }
